Rework. Keep `io.element.android.appconfig.LockScreenConfig` as simple as possible.
